pyproject.toml
[project]
name = "fine-tune-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bitsandbytes>=0.46.1",
    "datasets>=4.0.0",
    "docling>=2.43.0",
    "ollama>=0.5.1",
    "peft>=0.16.0",
    "python-dotenv>=1.1.1",
    "transformers>=4.53.3",
    "trl>=0.19.1",
]
